+++ crates/pipeline_manager/src/config.rs
@@ -4,8 +4,11 @@ use anyhow::{Error as AnyError, Result as AnyResult};
 use clap::Parser;
 use serde::Deserialize;
 use std::{
+    fmt::Display,
     fs::{canonicalize, create_dir_all},
     path::{Path, PathBuf},
+    str::FromStr,
+    string::ParseError,
 };
 
 const fn default_server_port() -> u16 {
@@ -230,6 +233,43 @@ impl ApiServerConfig {
     }
 }
 
+/// Argument to `cargo build --profile <>` passed to the rust compiler
+#[derive(Parser, Default, Deserialize, Debug, Clone)]
+pub enum CompilationProfile {
+    /// Prioritizes compilation speed over runtime speed
+    Unoptimized,
+    /// Prioritizes runtime speed over compilation speed
+    #[default]
+    Optimized,
+}
+
+impl FromStr for CompilationProfile {
+    type Err = ParseError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        match s {
+            "unoptimized" => Ok(CompilationProfile::Unoptimized),
+            "optimized" => Ok(CompilationProfile::Optimized),
+            e => unimplemented!(
+                "Unsupported option {e}. Available choices are 'unoptimized' and 'optimized'"
+            ),
+        }
+    }
+}
+
+impl Display for CompilationProfile {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match &self {
+            CompilationProfile::Unoptimized => write!(f, "unoptimized"),
+            CompilationProfile::Optimized => write!(f, "optimized"),
+        }
+    }
+}
+
+fn default_compilation_profile() -> CompilationProfile {
+    CompilationProfile::Optimized
+}
+
 /// Pipeline manager configuration read from a YAML config file or from command
 /// line arguments.
 #[derive(Parser, Deserialize, Debug, Clone)]
@@ -240,12 +280,15 @@ pub struct CompilerConfig {
     #[arg(long, default_value_t = default_working_directory())]
     pub compiler_working_directory: String,
 
-    /// Compile pipelines in debug mode.
-    ///
-    /// The default is `false`.
+    /// Pick the profile to use for cargo build.
+    /// Available choices are:
+    /// * 'unoptimized', for faster compilation times
+    /// at the cost of lower runtime performance.
+    /// * 'optimized', for faster runtime performance
+    /// at the cost of slower compilation times.
     #[serde(default)]
-    #[arg(long)]
-    pub debug: bool,
+    #[arg(long, default_value_t = default_compilation_profile())]
+    pub compilation_profile: CompilationProfile,
 
     /// Location of the SQL-to-DBSP compiler.
     #[serde(default = "default_sql_compiler_home")]
@@ -322,7 +365,7 @@ impl CompilerConfig {
     pub(crate) fn target_executable(&self, program_id: ProgramId) -> PathBuf {
         Path::new(&self.workspace_dir())
             .join("target")
-            .join(if self.debug { "debug" } else { "release" })
+            .join(self.compilation_profile.to_string())
             .join(Self::crate_name(program_id))
     }